requirements.txt
langchain==0.3.0
langchain-openai==0.2.0
langchain-community
streamlit==1.33.0
tika
faiss-cpu